Move `app.go` to pkg `application`
