+++ main.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"embed"
+	"gui-comicinfo/internal/application"
 
 	"github.com/wailsapp/wails/v2"
 	"github.com/wailsapp/wails/v2/pkg/options"
@@ -13,7 +14,7 @@ var assets embed.FS
 
 func main() {
 	// Create an instance of the app structure
-	app := NewApp()
+	app := application.NewApp()
 
 	// Create application with options
 	err := wails.Run(&options.App{
@@ -24,13 +25,13 @@ func main() {
 			Assets: assets,
 		},
 		BackgroundColour: &options.RGBA{R: 27, G: 38, B: 54, A: 1},
-		OnStartup:        app.startup,
+		OnStartup:        app.Startup,
 		Bind: []interface{}{
 			app,
 		},
 		EnumBind: []interface{}{
-			AllMangaValue,
-			AllAgeRatingValue,
+			application.AllMangaValue,
+			application.AllAgeRatingValue,
 		},
 	})
